added debts page
