+++ client/main.styl
@@ -2,6 +2,7 @@
 @import '../imports/ui/components/header/header.styl'
 @import '../imports/ui/components/footer/footer.styl'
 @import '../imports/ui/components/history/history.styl'
+@import '../imports/ui/components/debts/debts.styl'
 @import '../imports/ui/components/profile/profile.styl'
 @import '../imports/ui/components/feeds/feeds.styl'
 @import '../imports/ui/components/auth/auth.styl'
