+++ imports/ui/components/debts/debts.js
@@ -1,3 +1,22 @@
-/**
- * Created by denisbalyko on 07.05.16.
- */
+Template.debts.helpers({
+    debtsOrders: () => {
+        let orders = [];
+
+        Orders.find({ userId: Meteor.userId() }).fetch()
+            .forEach((baseOrder, index) => {
+                orders.push(baseOrder);
+
+                orders[index].poolOwnerId = Pools.findOne(baseOrder.poolId).ownerId;
+
+                baseOrder.items.forEach((baseItem, itemIndex) => {
+                    orders[index].items[itemIndex] = {
+                        id: baseItem.id,
+                        count: baseItem.count,
+                        item: Items.findOne({ _id: baseItem.id })
+                    };
+                });
+            });
+
+        return orders;
+    }
+});
