+++ imports/ui/components/header/header.html
@@ -16,7 +16,8 @@
                     {{#if currentUser}}
                     <ul class="nav navbar-nav">
                         <li><a href="{{pathFor 'pool_add'}}">Добавить пулл</a></li>
-                        <li><a href="{{pathFor 'history'}}">История</a></li>
+                        <li><a href="{{pathFor 'history'}}">Созданные пулы</a></li>
+                        <li><a href="{{pathFor 'debts'}}">Долги</a></li>
                         <li><a href="{{pathFor 'feeds'}}">Уведомления
                             {{#if feedsUnseenCount }}
                                 <span class="badge badge-inverse">{{ feedsUnseenCount }}</span>
