+++ imports/ui/components/history/history.js
@@ -1,20 +1,10 @@
 Template.history.helpers({
-    pools: Pools.find({ ownerId: Meteor.userId() }),
-    orders: () => {
-        let orders = [];
+    pools: () => {
+        return Pools.find({ ownerId: Meteor.userId() }).fetch().map((pool) => {
+            pool.poolPrice  = utils.getPriceWithFormat(Pools.getPoolPrice(pool._id));
+            pool.userCount = Orders.find({ poolId:pool._id }).count();
 
-        let baseOrders = Orders.find({ userId: Meteor.userId() }).fetch();
-        baseOrders.forEach((baseOrder, index) => {
-            orders.push(baseOrder);
-            baseOrder.items.forEach((baseItem, itemIndex) => {
-                orders[index].items[itemIndex] = {
-                    id: baseItem.id,
-                    count: baseItem.count,
-                    item: Items.findOne({ _id: baseItem.id })
-                };
-            });
+            return pool;
         });
-
-        return orders;
     }
 });
