+++ lib/router.js
@@ -28,7 +28,7 @@ Router.route('/profile', function() {
 
 Router.route('/history', function() {
     this.wait(Meteor.subscribe('PoolsListOwner', Meteor.userId()));
-    this.wait(Meteor.subscribe('OrdersListOwner', Meteor.userId()));
+    this.wait(Meteor.subscribe('OrdersCompany', Meteor.userId()));
     this.wait(Meteor.subscribe('OrdersItems', this.params.poolId));
     if (this.ready()) {
         if (Meteor.user()) {
@@ -40,6 +40,21 @@ Router.route('/history', function() {
         this.render('loading');
     }
 }, { name: 'history' });
+
+Router.route('/debts', function() {
+    this.wait(Meteor.subscribe('PoolsCompany', Meteor.user().profile.company));
+    this.wait(Meteor.subscribe('OrdersListOwner', Meteor.userId()));
+    this.wait(Meteor.subscribe('OrdersItems', this.params.poolId));
+    if (this.ready()) {
+        if (Meteor.user()) {
+            this.render('appDebts');
+        }else {
+            Router.go('/login');
+        }
+    }else {
+        this.render('loading');
+    }
+}, { name: 'debts' });
 
 Router.route('/feeds', function() {
     if (Meteor.user()) {
